+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 [...]
 
+- **[FIX]** Fix `Success Screen` gutter for large screens
+
 # v47.5.0 (12/02/2021)
 
 - **[FIX]** Fix `CheckCircleIcon` display when it has a title